asasdkiyadiygdqfiygtqwYUIDEQI

+++ src/gui/gui.cpp
@@ -6371,11 +6371,11 @@ bool FurnaceGUI::loop() {
             ImGuiIO& io = ImGui::GetIO();
             if(ImGui::Checkbox(id.c_str(),&pendingIns[i].second) && io.KeyShift)
             {
-              for(int jj = i - 1; jj >= 0; jj--)
+              for(int jj = (int)i - 1; jj >= 0; jj--)
               {
                 if(pendingIns[jj].second) //pressed shift and there's selected item above
                 {
-                  for(int k = jj; k < i; k++)
+                  for(int k = jj; k < (int)i; k++)
                   {
                     pendingIns[k].second = true;
                   }
